# Optimize searcher performance significantly