--- a/crates/maple_core/src/searcher/files.rs
+++ b/crates/maple_core/src/searcher/files.rs
@@ -6,7 +6,7 @@ use ignore::{DirEntry, WalkState};
 use matcher::Matcher;
 use printer::Printer;
 use std::path::PathBuf;
-use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::mpsc::{unbounded_channel, UnboundedSender};
@@ -17,7 +17,8 @@ fn search_files(
     hidden: bool,
     matcher: Matcher,
     stop_signal: Arc<AtomicBool>,
-    sender: UnboundedSender<Option<MatchedItem>>,
+    sender: UnboundedSender<MatchedItem>,
+    total_processed: Arc<AtomicUsize>,
 ) {
     let walk_config = WalkConfig {
         hidden,
@@ -31,14 +32,14 @@ fn search_files(
         let sender = sender.clone();
         let stop_signal = stop_signal.clone();
         let search_root = search_root.clone();
+        let total_processed = total_processed.clone();
         Box::new(move |entry: Result<DirEntry, ignore::Error>| -> WalkState {
             if stop_signal.load(Ordering::SeqCst) {
                 return WalkState::Quit;
             }
 
-            let entry = match entry {
-                Ok(entry) => entry,
-                Err(_) => return WalkState::Continue,
+            let Ok(entry) = entry else {
+                return WalkState::Continue;
             };
 
             // Only search file and skip everything else.
@@ -46,6 +47,8 @@ fn search_files(
                 Some(entry) if entry.is_file() => {}
                 _ => return WalkState::Continue,
             };
+
+            total_processed.fetch_add(1, Ordering::Relaxed);
 
             // TODO: Add match_file_path() in matcher to avoid allocation each time.
             let path = if let Ok(p) = entry.path().strip_prefix(&search_root) {
@@ -56,11 +59,15 @@ fn search_files(
 
             let maybe_matched_item = matcher.match_item(Arc::new(path));
 
-            if let Err(err) = sender.send(maybe_matched_item) {
-                tracing::debug!("Sender is dropped: {err:?}");
-                WalkState::Quit
-            } else {
-                WalkState::Continue
+            match maybe_matched_item {
+                Some(matched_item) => {
+                    if sender.send(matched_item).is_err() {
+                        WalkState::Quit
+                    } else {
+                        WalkState::Continue
+                    }
+                }
+                None => WalkState::Continue,
             }
         })
     });
@@ -81,38 +88,33 @@ pub async fn search(query: String, hidden: bool, matcher: Matcher, search_contex
 
     let (sender, mut receiver) = unbounded_channel();
 
-    std::thread::Builder::new()
-        .name("files-worker".into())
-        .spawn({
-            let stop_signal = stop_signal.clone();
-            move || search_files(paths, hidden, matcher, stop_signal, sender)
-        })
-        .expect("Failed to spawn blines worker thread");
+    let total_processed = Arc::new(AtomicUsize::new(0));
+
+    {
+        let total_processed = total_processed.clone();
+        std::thread::Builder::new()
+            .name("files-worker".into())
+            .spawn({
+                let stop_signal = stop_signal.clone();
+                move || search_files(paths, hidden, matcher, stop_signal, sender, total_processed)
+            })
+            .expect("Failed to spawn blines worker thread");
+    }
 
     let mut total_matched = 0usize;
-    let mut total_processed = 0usize;
 
     let printer = Printer::new(line_width, icon);
     let mut best_items = BestItems::new(printer, number, progressor, Duration::from_millis(200));
 
     let now = std::time::Instant::now();
 
-    while let Some(maybe_matched_item) = receiver.recv().await {
+    while let Some(matched_item) = receiver.recv().await {
         if stop_signal.load(Ordering::SeqCst) {
             return;
         }
-
-        match maybe_matched_item {
-            Some(matched_item) => {
-                total_matched += 1;
-                total_processed += 1;
-
-                best_items.on_new_match(matched_item, total_matched, total_processed);
-            }
-            None => {
-                total_processed += 1;
-            }
-        }
+        total_matched += 1;
+        let total_processed = total_processed.load(Ordering::Relaxed);
+        best_items.on_new_match(matched_item, total_matched, total_processed);
     }
 
     let elapsed = now.elapsed().as_millis();
@@ -125,6 +127,7 @@ pub async fn search(query: String, hidden: bool, matcher: Matcher, search_contex
     } = best_items;
 
     let display_lines = printer.to_display_lines(items);
+    let total_processed = total_processed.load(Ordering::SeqCst);
 
     progressor.on_finished(display_lines, total_matched, total_processed);