--- a/crates/maple_core/src/searcher/files.rs
+++ b/crates/maple_core/src/searcher/files.rs
@@ -117,6 +117,10 @@ pub async fn search(query: String, hidden: bool, matcher: Matcher, search_contex
         best_items.on_new_match(matched_item, total_matched, total_processed);
     }
 
+    if stop_signal.load(Ordering::SeqCst) {
+        return;
+    }
+
     let elapsed = now.elapsed().as_millis();
 
     let BestItems {