--- a/crates/maple_core/src/searcher/grep/stoppable_searcher.rs
+++ b/crates/maple_core/src/searcher/grep/stoppable_searcher.rs
@@ -6,15 +6,13 @@ use icon::Icon;
 use ignore::{DirEntry, WalkState};
 use matcher::Matcher;
 use std::path::PathBuf;
-use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 use tokio::sync::mpsc::{unbounded_channel, UnboundedSender};
 use types::{Rank, SearchProgressUpdate};
 
 pub(super) const UPDATE_INTERVAL: Duration = Duration::from_millis(200);
-
-pub(super) type SearcherMessage = crate::searcher::SearcherMessage<FileResult>;
 
 #[derive(Debug, Default)]
 struct MatchEverything;
@@ -52,7 +50,7 @@ pub struct FileResult {
 pub(super) struct StoppableSearchImpl {
     paths: Vec<PathBuf>,
     matcher: Matcher,
-    sender: UnboundedSender<SearcherMessage>,
+    sender: UnboundedSender<FileResult>,
     stop_signal: Arc<AtomicBool>,
 }
 
@@ -60,7 +58,7 @@ impl StoppableSearchImpl {
     pub(super) fn new(
         paths: Vec<PathBuf>,
         matcher: Matcher,
-        sender: UnboundedSender<SearcherMessage>,
+        sender: UnboundedSender<FileResult>,
         stop_signal: Arc<AtomicBool>,
     ) -> Self {
         Self {
@@ -71,7 +69,7 @@ impl StoppableSearchImpl {
         }
     }
 
-    pub(super) fn run(self) {
+    pub(super) fn run(self, total_processed: Arc<AtomicUsize>) {
         let Self {
             paths,
             matcher,
@@ -91,6 +89,7 @@ impl StoppableSearchImpl {
             let sender = sender.clone();
             let stop_signal = stop_signal.clone();
             let search_root = search_root.clone();
+            let total_processed = total_processed.clone();
             Box::new(move |entry: Result<DirEntry, ignore::Error>| -> WalkState {
                 if stop_signal.load(Ordering::SeqCst) {
                     return WalkState::Quit;
@@ -117,12 +116,14 @@ impl StoppableSearchImpl {
                     &MatchEverything,
                     entry.path(),
                     sinks::Lossy(|line_number, line| {
+                        total_processed.fetch_add(1, Ordering::Relaxed);
+
                         if line.is_empty() {
-                            // Discontinue if the sender has been dropped.
-                            return Ok(sender.send(SearcherMessage::ProcessedOne).is_ok());
+                            return Ok(true);
                         }
 
                         let line = line.trim();
+
                         let maybe_file_result =
                             matcher
                                 .match_file_result(relative_path, line)
@@ -135,14 +136,12 @@ impl StoppableSearchImpl {
                                     indices_in_line: matched.fuzzy_indices,
                                 });
 
-                        let searcher_message = if let Some(file_result) = maybe_file_result {
-                            SearcherMessage::Match(file_result)
-                        } else {
-                            SearcherMessage::ProcessedOne
-                        };
+                        if let Some(file_result) = maybe_file_result {
+                            // Discontinue if the sender has been dropped.
+                            return Ok(sender.send(file_result).is_ok());
+                        }
 
-                        // Discontinue if the sender has been dropped.
-                        Ok(sender.send(searcher_message).is_ok())
+                        Ok(true)
                     }),
                 );
 
@@ -200,16 +199,20 @@ pub async fn search(query: String, matcher: Matcher, search_context: SearchConte
 
     let (sender, mut receiver) = unbounded_channel();
 
+    let total_processed = Arc::new(AtomicUsize::new(0));
+
     std::thread::Builder::new()
         .name("grep-worker".into())
         .spawn({
             let stop_signal = stop_signal.clone();
-            move || StoppableSearchImpl::new(paths, matcher, sender, stop_signal).run()
+            let total_processed = total_processed.clone();
+            move || {
+                StoppableSearchImpl::new(paths, matcher, sender, stop_signal).run(total_processed)
+            }
         })
         .expect("Failed to spawn grep-worker thread");
 
     let mut total_matched = 0usize;
-    let mut total_processed = 0usize;
 
     let to_display_lines = |best_results: &[FileResult], icon: Icon| {
         let grep_results = best_results
@@ -261,77 +264,67 @@ pub async fn search(query: String, matcher: Matcher, search_context: SearchConte
     };
 
     let now = std::time::Instant::now();
-    while let Some(searcher_message) = receiver.recv().await {
+    while let Some(file_result) = receiver.recv().await {
         if stop_signal.load(Ordering::SeqCst) {
             return;
         }
 
-        match searcher_message {
-            SearcherMessage::Match(file_result) => {
-                total_matched += 1;
-                total_processed += 1;
+        total_matched += 1;
 
-                if best_results.results.len() <= best_results.max_capacity {
-                    best_results.results.push(file_result);
-                    best_results.sort();
+        let total_processed = total_processed.load(Ordering::Relaxed);
 
-                    let now = Instant::now();
-                    if now > best_results.past + UPDATE_INTERVAL {
-                        let display_lines = to_display_lines(&best_results.results, icon);
+        if best_results.results.len() <= best_results.max_capacity {
+            best_results.results.push(file_result);
+            best_results.sort();
+
+            let now = Instant::now();
+            if now > best_results.past + UPDATE_INTERVAL {
+                let display_lines = to_display_lines(&best_results.results, icon);
+                progressor.update_all(&display_lines, total_matched, total_processed);
+                best_results.last_lines = display_lines.lines;
+                best_results.past = now;
+            }
+        } else {
+            let last = best_results
+                .results
+                .last_mut()
+                .expect("Max capacity is non-zero; qed");
+
+            let new = file_result;
+            if new.rank > last.rank {
+                *last = new;
+                best_results.sort();
+            }
+
+            if total_matched % 16 == 0 || total_processed % 16 == 0 {
+                let now = Instant::now();
+                if now > best_results.past + UPDATE_INTERVAL {
+                    let display_lines = to_display_lines(&best_results.results, icon);
+
+                    let visible_highlights = display_lines
+                        .indices
+                        .iter()
+                        .map(|line_highlights| {
+                            line_highlights
+                                .iter()
+                                .copied()
+                                .filter(|&x| x <= line_width)
+                                .collect::<Vec<_>>()
+                        })
+                        .collect::<Vec<_>>();
+
+                    if best_results.last_lines != display_lines.lines.as_slice()
+                        || best_results.last_visible_highlights != visible_highlights
+                    {
                         progressor.update_all(&display_lines, total_matched, total_processed);
                         best_results.last_lines = display_lines.lines;
-                        best_results.past = now;
-                    }
-                } else {
-                    let last = best_results
-                        .results
-                        .last_mut()
-                        .expect("Max capacity is non-zero; qed");
-
-                    let new = file_result;
-                    if new.rank > last.rank {
-                        *last = new;
-                        best_results.sort();
+                        best_results.last_visible_highlights = visible_highlights;
+                    } else {
+                        progressor.quick_update(total_matched, total_processed)
                     }
 
-                    if total_matched % 16 == 0 || total_processed % 16 == 0 {
-                        let now = Instant::now();
-                        if now > best_results.past + UPDATE_INTERVAL {
-                            let display_lines = to_display_lines(&best_results.results, icon);
-
-                            let visible_highlights = display_lines
-                                .indices
-                                .iter()
-                                .map(|line_highlights| {
-                                    line_highlights
-                                        .iter()
-                                        .copied()
-                                        .filter(|&x| x <= line_width)
-                                        .collect::<Vec<_>>()
-                                })
-                                .collect::<Vec<_>>();
-
-                            if best_results.last_lines != display_lines.lines.as_slice()
-                                || best_results.last_visible_highlights != visible_highlights
-                            {
-                                progressor.update_all(
-                                    &display_lines,
-                                    total_matched,
-                                    total_processed,
-                                );
-                                best_results.last_lines = display_lines.lines;
-                                best_results.last_visible_highlights = visible_highlights;
-                            } else {
-                                progressor.quick_update(total_matched, total_processed)
-                            }
-
-                            best_results.past = now;
-                        }
-                    }
+                    best_results.past = now;
                 }
-            }
-            SearcherMessage::ProcessedOne => {
-                total_processed += 1;
             }
         }
     }
@@ -340,6 +333,7 @@ pub async fn search(query: String, matcher: Matcher, search_context: SearchConte
 
     let display_lines = to_display_lines(&best_results.results, icon);
 
+    let total_processed = total_processed.load(Ordering::SeqCst);
     progressor.on_finished(display_lines, total_matched, total_processed);
 
     tracing::debug!(