--- a/crates/maple_core/src/searcher/grep/stoppable_searcher.rs
+++ b/crates/maple_core/src/searcher/grep/stoppable_searcher.rs
@@ -159,6 +159,9 @@ impl StoppableSearchImpl {
 struct BestFileResults {
     /// Time of last notification.
     past: Instant,
+    // TODO: Use BinaryHeap than sorted vector.
+    //
+    // Blocked by https://github.com/rust-lang/rust/pull/124012
     results: Vec<FileResult>,
     last_lines: Vec<String>,
     last_visible_highlights: Vec<Vec<usize>>,