--- a/crates/maple_core/src/searcher/tagfiles.rs
+++ b/crates/maple_core/src/searcher/tagfiles.rs
@@ -348,6 +348,10 @@ pub async fn search(query: String, cwd: PathBuf, matcher: Matcher, search_contex
         best_items.on_new_match(matched_item, total_matched, total_processed);
     }
 
+    if stop_signal.load(Ordering::SeqCst) {
+        return;
+    }
+
     let elapsed = now.elapsed().as_millis();
 
     let BestItems {