This PR optimize the searcher performance even further, the hyperfine result indicates it's significantly improved by 60%. The series of optimization efforts have achieved 2x performance than v0.52.

```
hyperfine --warmup 3 './maple-v0.52 grep --ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client'  './maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client' './target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client'
Benchmark 1: ./maple-v0.52 grep --ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     657.0 ms ±  19.9 ms    [User: 1242.2 ms, System: 461.2 ms]
  Range (min … max):   610.5 ms … 677.2 ms    10 runs
 
Benchmark 2: ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     519.0 ms ±   9.2 ms    [User: 1014.1 ms, System: 354.4 ms]
  Range (min … max):   500.4 ms … 535.1 ms    10 runs
 
Benchmark 3: ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     322.0 ms ±   3.6 ms    [User: 514.1 ms, System: 215.0 ms]
  Range (min … max):   317.4 ms … 326.4 ms    10 runs
 
Summary
  ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client ran
    1.61 ± 0.03 times faster than ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
    2.04 ± 0.07 times faster than ./maple-v0.52 grep --ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
```